requirements.txt
tk 
pillow
requests
pynput
pyautogui
